Restrukturert litt kode
