+++ src/frontend/komponenter/Behandling/Vurdering/VurderingProps.ts
@@ -1,0 +1,12 @@
+import { IVilkårdata, IVurdering } from '../Inngangsvilkår/vilkår';
+import { Dispatch, ReactChild, SetStateAction } from 'react';
+import { IVilkårConfig } from './config/VurderingConfig';
+
+// Denne ligger separat slik att man kan gjennbruke den hvis man skal opprette flere typer av Vurderings-components
+export interface VurderingProps {
+    config: IVilkårConfig;
+    vilkårdata: IVilkårdata;
+    vurdering: IVurdering;
+    settVurdering: Dispatch<SetStateAction<IVurdering>>;
+    lagreKnapp: (visLagreKnapp: boolean) => ReactChild | undefined;
+}
