+++ src/frontend/komponenter/Behandling/Vurdering/VurderingUtil.ts
@@ -1,5 +1,5 @@
-import { IVurdering, Vilkårsresultat } from '../Inngangsvilkår/vilkår';
-import { VilkårGruppe, VurderingConfig } from './VurderingConfig';
+import { IVurdering, VilkårGruppe, Vilkårsresultat } from '../Inngangsvilkår/vilkår';
+import { VurderingConfig } from './config/VurderingConfig';
 
 export const alleErOppfylte = (vurderinger: IVurdering[]): boolean =>
     vurderinger.filter((vurdering) => vurdering.resultat !== Vilkårsresultat.JA).length === 0;
